requirements.txt
annotated-types==0.6.0  # Downgrade to support Python 3.12
invoice==1.0.1
invoke==2.2.0
kaggle==1.6.17
loguru==0.7.3
matplotlib==3.8.2
numpy==1.26.4
pandas==2.2.3
pydantic==2.10.5
pydantic-core>=2.0.2  # Allow pip to select a compatible version
pytest==6.2.4
safetensors==0.5.2
sympy==1.13.3
text-unidecode==1.3
tokenizers==0.21.0
toml==0.10.2
torch==2.2.0
tqdm==4.67.1
transformers==4.48.0
typer==0.15.1
typing_extensions==4.12.2
coverage==7.6.10
ruff
pre-commit
tabulate
fastapi
uvicorn
python-dotenv
httpx
praw
scikit-learn==1.6.1
hydra-core==1.3.2
